backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6

# RAG System
chromadb==0.4.22
sentence-transformers==2.3.1
langchain==0.1.6
langchain-community==0.0.19

# Document Processing
arxiv==2.1.0
biopython==1.83
pypdf2==3.0.1
pdfplumber==0.10.3

# LLM
requests==2.31.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
schedule==1.2.1
tqdm==4.66.1

# Export
python-docx==1.1.0
reportlab==4.0.9
jinja2==3.1.2
nbformat==5.9.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6

# Database
alembic==1.13.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
